+++ container-engine-lib/lib/backend_interface/objects/api_container/api_container.go
@@ -14,9 +14,39 @@ type APIContainer struct {
 
 	status container_status.ContainerStatus
 
+	// Private (i.e. internal to enclave) information about the API container
+	privateIpAddr net.IP
+	privateGrpcPort *port_spec.PortSpec
+	privateGrpcProxyPort *port_spec.PortSpec
+
 	// Public (i.e. external to Kurtosis) information about the API container
 	// This information will be nil if the API container isn't running
 	publicIpAddr net.IP
 	publicGrpcPort *port_spec.PortSpec
 	publicGrpcProxyPort *port_spec.PortSpec
+}
+
+func (apiContainer *APIContainer) GetEnclaveID() string {
+	return apiContainer.enclaveId
+}
+func (apiContainer *APIContainer) GetStatus() container_status.ContainerStatus {
+	return apiContainer.status
+}
+func (apiContainer *APIContainer) GetPrivateIPAddress() net.IP {
+	return apiContainer.privateIpAddr
+}
+func (apiContainer *APIContainer) GetPrivateGRPCPort() *port_spec.PortSpec {
+	return apiContainer.privateGrpcPort
+}
+func (apiContainer *APIContainer) GetPrivateGRPCProxyPort() *port_spec.PortSpec {
+	return apiContainer.privateGrpcPort
+}
+func (apiContainer *APIContainer) GetPublicIPAddress() net.IP {
+	return apiContainer.publicIpAddr
+}
+func (apiContainer *APIContainer) GetPublicGRPCPort() *port_spec.PortSpec {
+	return apiContainer.publicGrpcPort
+}
+func (apiContainer *APIContainer) GetPublicGRPCProxyPort() *port_spec.PortSpec {
+	return apiContainer.publicGrpcPort
 }